#refactoring and cleaning up
