+++ spark-ingest/config/README.md
@@ -1,0 +1,2 @@
+# heimdall-hoard
+A system that will ingest and hoard data from various data sources for use by search and recommendation systems.
